fix: add debug

+++ src/pages/dashboard/settings/Account.vue
@@ -148,16 +148,20 @@ async function deleteAccount() {
             return setErrors('update-account', [t('something-went-wrong-try-again-later')], {})
 
           try {
-            const userData = await supabaseClient
+            const { data: user } = await supabaseClient
               .from('users')
               .select()
               .eq('id', authUser.data.user.id)
               .single()
+            if (!user)
+              return setErrors('update-account', [t('something-went-wrong-try-again-later')], {})
 
-            await supabaseClient
-              .from('stripe_info')
-              .delete()
-              .eq('customer_id', userData.data?.customer_id)
+            if (user.customer_id) {
+              await supabaseClient
+                .from('stripe_info')
+                .delete()
+                .eq('customer_id', user.customer_id)
+            }
 
             const hashedEmail = await hashEmail(authUser.data.user.email!)
 
@@ -170,7 +174,7 @@ async function deleteAccount() {
             await supabaseClient
               .from('users')
               .delete()
-              .eq('id', userData.data?.id)
+              .eq('id', user.id)
 
             await deleteUser()
